server: Use more appropriate types for mk_server_worker_capacity

+++ src/include/mk_server.h
@@ -22,7 +22,7 @@
 #ifndef MK_SERVER_H
 #define MK_SERVER_H
 
-int mk_server_worker_capacity(int nworkers);
+unsigned int mk_server_worker_capacity(unsigned short nworkers);
 void mk_server_launch_workers(void);
 void mk_server_loop(int server_fd);
